requirements.txt
anyio==3.7.0
APScheduler==3.6.3
certifi==2023.5.7
cffi==1.15.1
cryptography==41.0.1
decorator==5.1.1
future==0.18.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
pycparser==2.21
python-telegram-bot==20.3
pytz==2023.3
six==1.16.0
sniffio==1.3.0
tornado==6.3.2
tzlocal==5.0.1
gspread==6.1.2
oauth2client==4.1.3